tests/test_personalization.py
import pytest
from frontend.src.contexts.PersonalizationContext import PersonalizationProvider, usePersonalization
from frontend.src.components.Personalization.PersonalizedContent import PersonalizedContent
from react import createElement
from react.testing import render, screen, act

# Note: These are conceptual tests as React testing in Python is complex
# In a real implementation, these would be Jest tests in the frontend

def test_personalization_context():
    """
    Test that the personalization context works correctly
    This is a conceptual test - actual implementation would use React testing library
    """
    # In a real test environment, we would:
    # 1. Render a component that uses the PersonalizationContext
    # 2. Verify that the experience level can be set and retrieved
    # 3. Verify that getPersonalizedContent returns appropriate content

    # For now, we'll just document the expected behavior:

    # When experience level is 'beginner', content for beginners should be returned
    # When experience level is 'advanced', content for advanced users should be returned
    # When specific level content is not available, it should fall back to the next appropriate level

    assert True  # Placeholder - actual tests would be in Jest for React components

def test_content_personalization_logic():
    """Test the logic for selecting personalized content"""
    # This would be a unit test for the getPersonalizedContent function
    # For now, we'll document the expected behavior:

    # If user is advanced and advanced content exists, return advanced content
    # If user is advanced but only beginner content exists, return beginner content
    # If user is intermediate and intermediate content exists, return intermediate content
    # etc.

    assert True  # Placeholder - actual tests would be in Jest for React components